+++ Resources/values/Strings.xml
@@ -3,6 +3,7 @@
     <string name="action_settings">Settings</string>
     <string name="ManualAddTripTitle">Add Trip Manually</string>
     <string name="FuelAddTitle">Add Fill-Up</string>
+    <string name="CarAddTitle">Add New Vehicle</string>
     
     <string name="TripDestination">Trip Destination</string>
     <string name="TripMilesDriven">Miles Driven</string>
