+++ UTLP.csproj
@@ -55,6 +55,7 @@
         <Reference Include="Mono.Android" />
     </ItemGroup>
     <ItemGroup>
+        <Compile Include="CarAddActivity.cs" />
         <Compile Include="FuelAddActivity.cs" />
         <Compile Include="FuelStruct.cs" />
         <Compile Include="MainActivity.cs" />
@@ -74,8 +75,15 @@
         <AndroidResource Include="Resources\layout\activity_main.axml">
             <SubType>Designer</SubType>
         </AndroidResource>
+        <AndroidResource Include="Resources\layout\activity_car_add.xml" />
         <AndroidResource Include="Resources\layout\activity_manual_trip_add.axml" />
+        <AndroidResource Include="Resources\layout\content_main.xml" />
+        <AndroidResource Include="Resources\menu\menu_car_add.xml">
+          <SubType>Designer</SubType>
+          <Generator>MSBuild:UpdateGeneratedFiles</Generator>
+        </AndroidResource>
         <AndroidResource Include="Resources\values\colors.xml" />
+        <AndroidResource Include="Resources\values\dimens.xml" />
         <AndroidResource Include="Resources\values\ic_launcher_background.xml" />
         <AndroidResource Include="Resources\values\strings.xml" />
         <AndroidResource Include="Resources\values\styles.xml" />
@@ -83,6 +91,7 @@
     <ItemGroup>
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
         <PackageReference Include="Xamarin.Android.Support.Design" Version="28.0.0.3" />
+        <PackageReference Include="Xamarin.Essentials" Version="1.5.3.2" />
     </ItemGroup>
     <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
     <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
